+++ out/rules/requireHrefOnAnchors.d.ts
@@ -1,2 +1,0 @@
-import { Rule } from '../linter';
-export default function requireHrefOnAnchors(): Rule;