+++ out/rules/requireHtmlLang.d.ts
@@ -1,2 +1,0 @@
-import { Rule } from '../linter';
-export default function requireHtmlLang(): Rule;